Import debugsubpkg.diff
We used to make a .debug symlink for each symlink whose target
has a .debug file to that file.  This is not necessary because
the debuglink section contains only the destination of those links.
Creating those links anyway results in debuginfo packages for
devel packages just because of the .so symlinks in them.